// frontend/workiez/src/models/Enums/ServicenameEnum.ts
export enum ServiceName {
    MASON = 'MASON', 
    ROOFER = 'ROOFER', 
    ELECTRICIAN = 'ELECTRICIAN', 
    PLUMBER = 'PLUMBER',
    HVAC = 'HVAC',
    COMPUTER_REPAIR = 'COMPUTER_REPAIR', 
    FURNITURE_REPAIR = 'FURNITURE_REPAIR', 
    TILE_SETTING = 'TILE_SETTING', 
    PLASTERING = 'PLASTERING',
    CLEANING = 'CLEANING', 
    TRANSPORT = 'TRANSPORT', 
    PAINTING = 'PAINTING', 
    WASTE_AND_ENVIRONMENTAL_SERVICES = 'WASTE_AND_ENVIRONMENTAL_SERVICES',
    PEST_CONTROL = 'PEST_CONTROL', 
    CARPENTERS = 'CARPENTERS'
}